autoload for appenders
