+++ lib/logging/appenders/growl.rb
@@ -1,13 +1,6 @@
 
 module Logging::Appenders
 
-  # Accessor / Factory for the Growl appender.
-  #
-  def self.growl( *args )
-    return ::Logging::Appenders::Growl if args.empty?
-    ::Logging::Appenders::Growl.new(*args)
-  end
-
   # This class provides an Appender that can send notifications to the Growl
   # notification system on Mac OS X.
   #
