+++ lib/logging/appenders/io.rb
@@ -1,13 +1,6 @@
 
 module Logging::Appenders
 
-  # Accessor / Factory for the IO appender.
-  #
-  def self.io( *args )
-    return ::Logging::Appenders::IO if args.empty?
-    ::Logging::Appenders::IO.new(*args)
-  end
-
   # This class provides an Appender that can write to any IO stream
   # configured for writing.
   #
