+++ lib/logging/appenders/syslog.rb
@@ -6,13 +6,6 @@ if HAVE_SYSLOG
 
 module Logging::Appenders
 
-  # Accessor / Factory for the Syslog appender.
-  #
-  def self.syslog( *args )
-    return ::Logging::Appenders::Syslog if args.empty?
-    ::Logging::Appenders::Syslog.new(*args)
-  end
-
   # This class provides an Appender that can write to the UNIX syslog
   # daemon.
   #
